refactor: split block generator service helpers
